Backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1